Split some tasks out of mw.c into fc_tasks.c
